pyproject.toml
[tool.poetry]
name = "dave"
version = "2.0.3"
description = "DAVE proxy for the Crystals GFlowNet - Conventional cell @ MatBench"
authors = [
    "vict0rsch <vsch@pm.me>",
    "AlexDuvalinho <alexandre.duval@mila.quebec>",
    "sh-divya <divya.sharma@mila.quebec>",
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = ">=2.0.1"
torch-geometric = ">=2.3.1"
faenet = ">=0.1.3"
phast = ">=0.1.2"
pyxtal = ">=0.6.0"
pymatgen = ">=2023.8.10"
lightning = ">=2.0.8"
scikit-learn = ">=1.3.0"
numpy = ">=1.21"
matplotlib = ">=3.7.3"
click = ">=8.1.7"
ase = ">=3.22.1"
pandas = ">=2.0.0"
mendeleev = ">=0.14.0"
tqdm = ">=4.66.1"
scipy = ">=1.11.2"
PyYAML = ">=6.0.1"
minydra = ">=0.1.6"
rich = ">=13.5.2"
lmdb = ">=1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
